Created new API Interface File
- Created api_v1_0
- Began Profiling
- prepared functions to be affected by timestamp changes
