+++ plotting/plotter.py
@@ -16,6 +16,7 @@ from oceannavigator.dataset_config import (
 from flask_babel import format_date, format_datetime
 import contextlib
 from PIL import Image
+from utils.function_profiler import profileit
 
 # Base class for all plotting objects
 class Plotter(metaclass=ABCMeta):
@@ -44,6 +45,7 @@ class Plotter(metaclass=ABCMeta):
 
     # Called by views.py to parse query, load data, and return the generated file
     # to be displayed by Javascript.
+    @profileit
     def run(self, **kwargs):
         if 'size' in kwargs and kwargs.get('size') is not None:
             self.size = kwargs.get('size')
